% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluxSeparatorApp.R
\name{FluxSeparatorApp}
\alias{FluxSeparatorApp}
\title{FluxSeparator Shiny App}
\usage{
FluxSeparatorApp(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny]{shinyApp}()}.}
}
\description{
Shiny app to visualize the process of separating diffusive and ebullitive fluxes. The app also gives the user the ability to
mark areas where diffusive fluxes should be calculated.
}
\examples{
\donttest{
library(FluxSeparator)

FluxSeparatorApp()
}
}
\references{
Sø et al. (2024). Self-Made Equipment for Automatic Methane
Diffusion and Ebullition Measurements From Aquatic Environments.
\doi{10.1029/2024JG008035}.

Sø et al. (2023). Methane and carbon dioxide fluxes at high
spatiotemporal resolution from a small temperate lake.
\doi{10.1016/j.scitotenv.2023.162895}.

Hutchinson, G.L. and Mosier, A.R. (1981). Improved soil cover
method for field measurement of nitrous oxide fluxes. Soil Science Society
of America Journal, 45, pp. 311-316.

Pullens, J.W.M., Abalos, D., Petersen, S.O. and Pedersen, A.R.
(2023). Identifying criteria for greenhouse gas flux estimation with
automatic and manual chambers: A case study for N2O. European Journal of
Soil Science, 74, e13340. \doi{10.1111/ejss.13340}.
}
\seealso{
\code{\link{diffusive_flux}}, \code{\link{ebullitive_flux}}, \code{\link{ppm_to_umol}}
}
\author{
Jonas Stage Sø \email{Jonassoe@biology.sdu.dk}
}
