% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_to_umol.R
\name{ppm_to_umol}
\alias{ppm_to_umol}
\title{ppm_to_µmol}
\usage{
ppm_to_umol(pressure, concentration, volume, temperature_C, area)
}
\arguments{
\item{pressure}{Air pressure during measurement (Pa).}

\item{concentration}{Concentration of the gas in ppm (µmol/mol).}

\item{volume}{Volume of the chamber used for measuring in \eqn{m^{3}}.}

\item{temperature_C}{Temperature in degrees Celsius in the chamber.}

\item{area}{Surface area of the chamber used in \eqn{m^{2}}.}
}
\value{
A numeric vector of flux values in \eqn{\mu mol} \eqn{m^{-2}} \eqn{h^{-1}}.
}
\description{
Conversion of \eqn{ppmV} to \eqn{\mu mol} \eqn{m^{-2}} \eqn{h^{-1}} using the ideal gas law.
}
\examples{
# Convert a single value
ppm_to_umol(pressure = 101325, concentration = 10,
            volume = 0.01, temperature_C = 20, area = 0.05)

}
\author{
Jonas Stage Sø \email{Jonassoe@biology.sdu.dk}
}
