% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIY_sensor_data.R
\docType{data}
\name{DIY_sensor_data}
\alias{DIY_sensor_data}
\title{Data from Lake Lyng}
\source{
{Load a subset of the dataset.}
}
\arguments{
\item{datetime}{Date and time of measurement}

\item{RH\%}{Relative humidity}

\item{tempC}{Temperature in degree celcius}

\item{CH4smV}{Methane sensor voltage}

\item{K33_RH}{Relative humidity measured by the CO2 sensor}

\item{K33_Temp}{Temperature in degree celcius measured by the CO2 sensor}

\item{K33_CO2}{CO2 concentration in ppm}

\item{SampleNumber}{Sample number in this pump cycle}

\item{PumpCycle}{Pump cycle which counts upwards after the chamber has been flushed}

\item{pred_CH4}{Predicted methane concentration, this is automaticly calculated using the read_CH4_files function, and calibration values}

\item{station}{Station name. Needed for diffusive and ebullitive fluxes calculations}

\item{sensor}{Sensor name}
}
\value{
A data frame with 2,201 rows and 10 variables.
}
\description{
A subset of data from the paper Methane and carbon dioxide fluxes at high spatiotemporal resolution from a small temperate lake.
One measurement containing only diffusive flux and one containing ebullitive events
}
\examples{
data(DIY_sensor_data)
}
\references{
Sø et al. (2023). Methane and carbon dioxide fluxes at high spatiotemporal resolution from a small temperate lake. DOI: \if{html}{\out{<a href="http://dx.doi.org/10.1016/j.scitotenv.2023.162895">}}http://dx.doi.org/10.1016/j.scitotenv.2023.162895\if{html}{\out{</a>}}.
}
\author{
Jonas Stage Sø \email{Jonassoe@biology.sdu.dk}
}
