% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_to_umol.R
\name{ppm_to_umol}
\alias{ppm_to_umol}
\title{ppm_to_µmol}
\usage{
ppm_to_umol(pressure, concentration, volume, temperature_C, area)
}
\arguments{
\item{pressure}{Air pressure during measurement (Pa).}

\item{concentration}{Concentration of the gas in ppm (µmol/mol).}

\item{volume}{Volume of the chamber used for measuring in \eqn{m^{3}}.}

\item{temperature_C}{Temperature in degrees celcius in the chamber.}

\item{area}{Surface area of the chamber used in \eqn{m^{2}}}
}
\description{
Convertion of \eqn{ppmV} to \eqn{µmol} \eqn{m^{-2}} \eqn{h^{-1}} using the ideal gas law.
}
\examples{
\dontrun{
library(FluxSeparator)

# See ?ebullitive_flux or ?diffusive_flux for help with the respective functions

# Using the ebullitive flux function
DIY_sensor_data \%>\%
  ebullitive_flux() \%>\%
  mutate(umol_m2_h1_flux = ppm_to_umol(pressure
                                       concentration
                                       volume
                                       temperature_C
                                       area))

# Using the diffusive flux function
DIY_sensor_data \%>\%
 diffusive_flux(cutoff_start_value = 5) \%>\%
# 450 would be good for CO2, while 5 could be good for CH4
 mutate(umol_m2_h1_flux = ppm_to_umol(pressure
                                      concentration
                                      volume
                                      temperature_C
                                      area))
}
}
\author{
Jonas Stage Sø \email{Jonassoe@biology.sdu.dk}
}
